# psk-ui/src/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenApi specification - Hasan Burak Songur",
    "description": "OpenApi documentation for Spring Security",
    "termsOfService": "Terms of service",
    "contact": {
      "name": "Hasan Burak Songur",
      "url": "hasanburaksongur.com",
      "email": "hasanburaksongur@gmail.com"
    },
    "license": {
      "name": "Licence name",
      "url": "https://some-url.com"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8088/api",
      "description": "Local ENV"
    },
    {
      "url": "https://hasanburaksongur.com",
      "description": "PROD ENV"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/user/user": {
      "get": {
        "tags": ["User"],
        "operationId": "getUser",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["User"],
        "operationId": "updateUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v2/sessions/update-status": {
      "post": {
        "tags": ["session-controller-v-2"],
        "operationId": "updateSessionStatusV2",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/v2/sessions/update-paid-status": {
      "post": {
        "tags": ["session-controller-v-2"],
        "operationId": "updateSessionPaidStatusV2",
        "parameters": [
          {
            "name": "isPaid",
            "in": "query",
            "required": true,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "sessionId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponseV2"
                }
              }
            }
          }
        }
      }
    },
    "/v2/sessions/update-date": {
      "post": {
        "tags": ["session-controller-v-2"],
        "operationId": "updateSessionDateV2",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionDateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/v2/sessions/make-unavailable": {
      "post": {
        "tags": ["session-controller-v-2"],
        "operationId": "makeUnavailableV2",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string",
                  "format": "date-time"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PublicSessionResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v2/sessions/make-available": {
      "post": {
        "tags": ["session-controller-v-2"],
        "operationId": "makeAvailableV2",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string",
                  "format": "date-time"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PublicSessionResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v2/sessions/create": {
      "post": {
        "tags": ["session-controller-v-2"],
        "operationId": "createSessionForUserV2",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/v2/sessions/add-note/user": {
      "post": {
        "tags": ["session-controller-v-2"],
        "operationId": "addNoteToSessionForUserV2",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionUserNoteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponseV2"
                }
              }
            }
          }
        }
      }
    },
    "/v2/sessions/add-note/psychologist": {
      "post": {
        "tags": ["session-controller-v-2"],
        "operationId": "addNoteToSessionForPsychologistV2",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionPsychologistNoteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponseV2"
                }
              }
            }
          }
        }
      }
    },
...
    "/sessions/{id}/cancel": {
      "post": {
        "tags": ["session-controller"],
        "operationId": "cancelUserSession",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
    "/sessions/create": {
      "post": {
        "tags": ["session-controller"],
        "operationId": "createSessionForUserV2_1",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
...
    "/v2/sessions/users-with-sessions": {
      "get": {
        "tags": ["session-controller-v-2"],
        "operationId": "getAllUsersWithSessionV2",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseUserWithIncomingSessionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v2/sessions/user/{userId}": {
      "get": {
        "tags": ["session-controller-v-2"],
        "operationId": "getAllSessionsOfUserV2",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SessionResponseV2"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/v2/sessions/all": {
      "get": {
        "tags": ["session-controller-v-2"],
        "operationId": "getAllSessionsV2",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseSessionResponseV2"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "UserRequest": {
        "type": "object",
        "properties": {
          "firstname": {
            "type": "string"
          },
          "lastname": {
            "type": "string"
          },
          "phoneNumber": {
            "maxLength": 10,
            "minLength": 10,
            "type": "string"
          },
          "birthYear": {
            "maxLength": 4,
            "minLength": 4,
            "type": "string"
          }
        }
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "firstname": {
            "type": "string"
          },
          "lastname": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "birthYear": {
            "type": "string"
          }
        }
      },
...
      "SessionStatusRequest": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "integer",
            "format": "int32"
          },
          "sessionStatusType": {
            "type": "string",
            "enum": [
              "AWAITING_PSYCHOLOGIST_APPROVAL",
              "AWAITING_PAYMENT",
              "AWAITING_PAYMENT_CONFIRMATION",
              "APPOINTMENT_SCHEDULED",
              "COMPLETED",
              "CANCELED",
              "UNAVAILABLE"
            ]
          }
        }
      },
      "SessionResponseV2": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "integer",
            "format": "int32"
          },
          "userForSessionResponse": {
            "$ref": "#/components/schemas/UserForSessionResponse"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "sessionStatus": {
            "type": "string",
            "enum": [
              "AWAITING_PSYCHOLOGIST_APPROVAL",
              "AWAITING_PAYMENT",
              "AWAITING_PAYMENT_CONFIRMATION",
              "APPOINTMENT_SCHEDULED",
              "COMPLETED",
              "CANCELED",
              "UNAVAILABLE"
            ]
          },
          "noteForUser": {
            "type": "string"
          },
          "noteForPsychologist": {
            "type": "string"
          },
          "isPaid": {
            "type": "boolean"
          }
        }
      },
      "UserForSessionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "firstname": {
            "type": "string"
          },
          "lastname": {
            "type": "string"
          }
        }
      },
      "SessionDateRequest": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "integer",
            "format": "int32"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "PublicSessionResponse": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "SessionUserNoteRequest": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "integer",
            "format": "int32"
          },
          "noteForUser": {
            "type": "string"
          }
        }
      },
      "SessionPsychologistNoteRequest": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "integer",
            "format": "int32"
          },
          "noteForPsychologist": {
            "type": "string"
          }
        }
      },
...
      "PageResponseUserWithIncomingSessionResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserWithIncomingSessionResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      },
      "UserWithIncomingSessionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "firstname": {
            "type": "string"
          },
          "lastname": {
            "type": "string"
          },
          "sessionResponse": {
            "$ref": "#/components/schemas/SessionResponseV2"
          }
        }
      },
      "PageResponseSessionResponseV2": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SessionResponseV2"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT auth description",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}